Implemented Feature#55 - "for" parameter for mostpopular action.
The "for" argument lets you pick a particular page to start the popularity list from in the page tree.

The "dontrecurse" argument tells it to only include the direct children of the specified page.  If it is excluded or set to false it will show all children, grandchildren etc for that page.  This is the default behavior.