Implemented Feature#102 - PageIndex by letter including result limiting.
Added two arguments to the PageIndex action.  The first is "limit" which sets the maximum number of pages to show at any one time on each index page.  The second argument is "letter" which lets you pick which pages to display based on their name starting with a specific letter.